Update SDK to version 25.06.0 for iMXRT11xx

+++ boards/evkbmimxrt1170/audio_examples/maestro_usb_mic/cm7/hardware_init.c
@@ -122,7 +122,7 @@ void BOARD_InitHardware(void)
 
     /* Init DMAMUX */
     DMAMUX_Init(DEMO_DMAMUX);
-    DMAMUX_SetSource(DEMO_DMAMUX, DEMO_DMA_RX_CHANNEL, (uint8_t)DEMO_SAI_RX_SOURCE);
+    DMAMUX_SetSource(DEMO_DMAMUX, DEMO_DMA_RX_CHANNEL, DEMO_SAI_RX_SOURCE);
     DMAMUX_EnableChannel(DEMO_DMAMUX, DEMO_DMA_RX_CHANNEL);
 
     /* Clock setting for USB */
