+++ boards/evkmimxrt1160/driver_examples/sai/edma_transfer/cm7/hardware_init.c
@@ -81,7 +81,7 @@ void BOARD_InitHardware(void)
 
     /* Init DMAMUX */
     DMAMUX_Init(DEMO_DMAMUX);
-    DMAMUX_SetSource(DEMO_DMAMUX, DEMO_EDMA_CHANNEL, (uint8_t)DEMO_SAI_TX_SOURCE);
+    DMAMUX_SetSource(DEMO_DMAMUX, DEMO_EDMA_CHANNEL, DEMO_SAI_TX_SOURCE);
     DMAMUX_EnableChannel(DEMO_DMAMUX, DEMO_EDMA_CHANNEL);
 }
 /*${function:end}*/
